Track slashes, methods, and events"
